[src] Move definition of DEFAULT_CHUNK_SIZE
