+++ audyn/bin/download_lsx.py
@@ -7,9 +7,8 @@ import zipfile
 
 from omegaconf import DictConfig
 
-from ..utils._download import DEFAULT_CHUNK_SIZE
 from ..utils._hydra import main as audyn_main
-from ..utils.data.download import download_file
+from ..utils.data.download import DEFAULT_CHUNK_SIZE, download_file
 
 
 @audyn_main(config_name="download-lsx")
