+++ audyn/bin/download_musdb18.py
@@ -7,9 +7,8 @@ from urllib.request import Request, urlopen
 
 from omegaconf import DictConfig
 
-from ..utils._download import DEFAULT_CHUNK_SIZE
 from ..utils._hydra import main as audyn_main
-from ..utils.data.download import download_by_response
+from ..utils.data.download import DEFAULT_CHUNK_SIZE, download_by_response
 
 try:
     from tqdm import tqdm
