--- audyn/utils/_download/__init__.py
@@ -1,5 +1,0 @@
-__all__ = [
-    "DEFAULT_CHUNK_SIZE",
-]
-
-DEFAULT_CHUNK_SIZE = 8192
